[changed] TestLocation is a constructor
TestLocation is no longer a singleton. Instead, it's a constructor.
This has the side effect that when a new TestLocation is created it
automatically updates History.length. In practice this shouldn't be
a problem any more than the singleton implementation was, since it
still permits you to run only one test at a time.