added `createElement` prop to `Router`
This gives control of rendering the route components to the consumer.

Particularly useful for data abstractions like Relay/Falkor/AsyncProps,
where components need to be rendered by a container component that uses
static properties of the component to load data and subscribes to the
data stores, re-rendering the route component when data updates, instead
of only when the routes change.