expose a Route's match through HOCs
The `matchProvider` HOC wraps the component that will be rendered by a `<Route>` and makes the `match` available through the context. The `withMatch` HOC wraps a component that wants access to its parent `<Route>`'s match.

This approach requires the user to specify that they want the match made available through the context, and could lead to issues with users that neglect to wrap a nested `<Route>`'s component in a `matchProvider` and end up accessing the match of a grandparent `<Route>`. An alternative to this would be for either the `Route` to expose the context variable (it currently cannot because it is a functional component) or for `createRouteElement` to automatically wrap what it renders in a `<MatchProvider match={props.match}>` component.