Treat all navigation as a side effect
- Removes <Redirect>
- Removes support for { redirectTo } in useRoutes()
- Warns if navigate() is used during the initial render
- Warns when <Navigate> is used on the initial render in <StaticRouter>

This began as a discussion on GitHub (see #7203). Essentially we need to
treat all navigation as a side effect, which means it needs to happen in
a useEffect. If people need v5's <Redirect> functionality (when it was
used in the route config inside a <Switch>), they either need to a) do
the redirect on the server (the best solution) or b) render a <Navigate>
on the initial render. However, (b) will only show the new route on the
subsequent render.

This should also eliminate warnings in the dev-experimental builds.

Closes #7203
Fixes #7199