Fix greedy * in route paths
This prevents `*` from matching without a preceding `/` in the URL
pathname. So e.g. `/users/*` matches `/users/mj` but not `/userstypo`.

This also fixes `to="."` in splat routes so they point to the route
path *including* the portion of the URL that was matched by the `*`,
which also makes `*` more consistent with `:param` since it's treated
just the same as any other param.

There is however a subtle breaking change if you are using the low-level
`match` API, e.g. the `match` object you get back from `matchRoutes()`
or `useMatch()`. `match.pathname` in a splat route now includes the full
URL pathname that was matched instead of only the portion before the
`*`. There is a new variable, `match.pathnameStart` that you can use if
you needed this for doing your own route matching.

Fixes #7972